refactor: reorganize folders
