+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Layout from './common/NavBar';
-import Login from './pages/login/Login';
+import Login from './pages/Login';
 import Registration from './pages/Registration';
 import Main from './pages/Main';
 import NotFound from './pages/NotFound';
